Split single bats test into individual jobs per test file (#83)
Splitting the tests into multiple jobs allows them to run in parallel, which cuts the overall runtime from around 25 minutes to 12 minutes with full parallelism.

We could've done that in a single matrix, but that would've lead to a long list of individual jobs. Extracting a workflow call makes it more structured by machine.